merge

+++ ogn/src/DB/ConfigManager.cpp
@@ -224,12 +224,12 @@ void ConfigManager::loadPropertyJson(const std::string& path)
 
 		PropertyJson propertyJson;
 		propertyJson.ID = v["ID"].asUInt();
-		propertyJson.Speed = v["Speed"].asDouble();
+		propertyJson.Speed = v["Speed"].asFloat();
 		propertyJson.MaxHp = v["MaxHp"].asUInt();
 		propertyJson.MaxMp = v["MaxMp"].asUInt();
 		propertyJson.Attack = v["Attack"].asUInt();
 		propertyJson.Defense = v["Defense"].asUInt();
-		propertyJson.SpeedUp = v["SpeedUp"].asDouble();
+		propertyJson.SpeedUp = v["SpeedUp"].asFloat();
 		propertyJson.MaxHpUp = v["MaxHpUp"].asUInt();
 		propertyJson.MaxMpUp = v["MaxMpUp"].asUInt();
 		propertyJson.AttackUp = v["AttackUp"].asUInt();
